requirements.txt
# basic
numpy~=2.2.5
pandas~=2.2.3
scipy
scikit-learn~=1.6.1
tqdm~=4.67.1

# notebooks
jupyterlab
ipywidgets
notebook
jupyterlab-widgets

matplotlib~=3.10.3
tabulate~=0.9.0
pip~=24.3.1
attrs~=25.3.0
pillow~=11.2.1
tornado~=6.4.2
Jinja2~=3.1.6
filelock~=3.18.0
opencv-python~=4.11.0.86
pytesseract~=0.3.13
torch~=2.7.0
requests~=2.32.3
transformers~=4.51.3
termcolor~=3.1.0
ipython~=9.2.0
evaluate~=0.4.3
pytorch-lightning~=2.5.1.post0

Flask>=3.0.0
Flask-CORS>=4.0.0
Pillow>=10.0.0
Werkzeug>=3.0.0
pandas>=1.0
numpy>=1.18
matplotlib
scipy
tabulate